+++ src/functions/implements/math/exp.c
@@ -84,7 +84,7 @@ rt_function_error_t exec_exp(rt_function_t *f) {
 
   int i; // Iterator
   for (i = 0; i < p->output_size; i++) {
-    y[i] = (float)exp(x[i]);
+    y[i] = expf(x[i]);
   }
   return RT_FUNCTION_ERROR_NOERROR;
 }
@@ -97,7 +97,7 @@ rt_function_error_t exec_exp_generic(rt_function_t *f) {
   int i; // Iterator
   for (i = 0; i < p->output_size; i++) {
     float x = p->get_input(p->input, i);
-    p->set_output(p->output, i, (float)exp(x));
+    p->set_output(p->output, i, expf(x));
   }
   return RT_FUNCTION_ERROR_NOERROR;
 }
